refactoring lsstsw tutorial
This tutorial was originally written before the Workflow page in the DM
developer guide. There was a lot of overlap in content once that
workflow document was written. However the workflow document is intended
to capture official processes, while this lsstsw tutorial has always
been intended to be hands-on. Thus this commit removes policy definition
from the lsstsw tutorial and links heavily to the DM workflow
documentation.

For DM-4793.