Refactor release notes to running page
Release notes are no longer managed on separate pages, but instead are
collected on a single page at releases/release-notes.

Similarly the known issues are also collected on a single page.

The idea is that these pages will reflect the full state/history up to
the release point.

However the metric characterization reports are still maintained in
separate pages since these might be interest in viewing older reports
even from a newer release. We may revisit this policy.