Move release notes to releases/notes.rst
This strategy is to make way for release announcements, which will also
be contained in the releases/ directory, but in individual pages.

The releases/index.rst page is now 'hidden' in the toctree since the
intention is for a user to only see the announcement page for the current
release, and to also see the collected release notes directly from the
main page.