Use LTD edition names + CURRENT_RELEASE config var
In conf.py, the new CURRENT_RELEASE variable is where the last stable
release is document. That's considered the version when building master.

For ticket branches, we use the form_ltd_edition_name() function
in documenteer to build edition names that mirror LSST the Docs edition
names. For example, a tickets/DM-9686 branch becomes just DM-9686.