Re-organize homepage to feature getting-started
The idea is that the getting started section is special: it has
tutorials designed for the best initial onboarding experience.

We might want to add an overview document that explains what the science
pipelines are and what features they provide.

We deliberately separate the getting started from the installation
sections. This lets us use the installation section to provide more
detailed installation instructions for different use-cases.

The Getting Started pages guide a user through that installation
documentation.