Add display tutorial as Part3 of Getting Started
This tutorial covers the display framework for visualizing source images
and catalogs. The tutorial is part of the Getting Started series to
introduce readers to the software.

I've inserted the display tutorial after the processCcd.py tutorial so
we can view images and source catalogs from processCcd.py before going
on to coaddition.

The commit also updates the tutorial numbering sequence.