Document the need for flock.
We require that filesystems used for output of certain files (e.g. persisted configurations, schemas, etc.) support the flock system call.  Tests of this capability in `daf_persistence` also require that the build filesystem have the same support.  There does not appear to be a good way to check for filesystem support, so for now documenting this requirement here seems like the best that can be done.  Also suggest how flock support might be manually checked.