Update docker.rst
`/home/lsst` is the default user home directory in Docker image `w_2018_06`.  I haven't checked other images.

This line should be changed.  Should the examples be changed to be with respect to `/home/lsst`?

Also, the container opens up in `/opt/lsst/software/stack`.  This is somewhat convenient as you don't have to look up the path for the `loadLSST.bash` script, but perhaps a note to this effect that the containers don't open in the home directory will help the naive user who is completely lost.