Create a new Tickets page to gather all tickets
We started gather a list of tickets for each release. The trouble is
that we added these a toctree items under "Release History" so that the
Table of Contents would look like this:

- LSST Stack v15.0 Release
- LSST Stack v14.0 Release
- LSST Stack v13.0 Release
- LSST Stack v12.0 Release
- Release Notes
- Tickets Addressed in Spring 2018 Release (v15_0)
- Tickets Addressed in Spring 2017 Release (v14_0)

This PR changes that table of contents to:

- LSST Stack v15.0 Release
- LSST Stack v14.0 Release
- LSST Stack v13.0 Release
- LSST Stack v12.0 Release
- Release Notes
- Tickets

Similar to Release Notes, Tickets gathers reference content from all
releases. This table of contents still has some weirdness (why are there
multiple announcements? when the old releases are also announced in the
release notes?), but overall it's a more scalable approach.