Streamline contrib info in README
It turned out that linking to the Jenkins job in the README, it enticed
a developer to try running that Jenkins job without going through the
documentation to learn how to why to use it, creating a lot of
confusion.

Now the README focuses on just saying that the docs are built
automatically, and then links to the CONTRIBUTING file that can more
carefully guide developers through how to build the documentation.