Template version strings in installation docs
Now the installation docs use the jinja directive (available in
documenteer 0.5.0+) to template the EUPS tags and other version-specific
information. This version information is generated from documenteer's
configuration code, and ultimately is based on the EUPS_TAG environment
variable that exists in the build environment.

- Git ref of the newinstall.sh file
- EUPS tag for the eups distrib install command
- Git ref of the demo
- Docker tags
- Links to conda environment files

Overall this means that daily/weekly/major builds have installation
documentation that reflects the version of the software being built.