Update release notes for 23.0.1
Updated existing 23.0.0 page rather than making a new one
because 23.0.1 is a patch for 23.0.0. We do not expect or
recommend than anyone use 23.0.0 for any reason. For
reproducing DP0.2, the results of single frame processing
will be the same with 23.0.1 and 23.0.0, as this patch includes
no behavior changes to single frame processing.