Fix plot extension patch for documenteer 0.6
This patches the preset configuration of the matplotlib plot directive
from documenteer. In documenteer 0.5 we automatically configure the docs
with this extension, but in documenteer 0.6 we stopped adding this
extension by default because of compatibility issues with the latest
Sphinx. Once we move the Pipelines docs build in ci.lsst.codes to
Documenteer 0.6+ we can drop this patch.