Document LSST_CONDA_ENV_NAME envvar and add link for versions.
This is the same blurb in three places, but I think that's fine, as we
don't want users to have to jump around to find this information any
more than we wanted them to jump around to find what's already
duplicated on these pages.

It might also belong in the lsstsw instructions, but I wasn't sure it
would actually work there.